Configuration Time Constants (#19)
You may notice that config time constants were already added to the language (#2549). This implementation did not actually address the use cases it was intended for, though, and so this is a request for comments on v2 of the notion. The new implementation of configuration time constants will be more complicated as it involves interactions between the SDK and Sway via a new descriptor file, but it does maintain independence between the two tools so updating constants won't require a recompile.
---------

Co-authored-by: Alex <alex@system76-pc.localdomain>
Co-authored-by: Mazdak Farrokhzad <twingoow@gmail.com>
Co-authored-by: Mohammad Fawaz <mohammadfawaz89@gmail.com>
Co-authored-by: mitchmindtree <mail@mitchellnordine.com>
Co-authored-by: iqdecay <victor@berasconsulting.com>